fix: keep activity field helpers off the DB import path
Unit tests for pure field-change helpers were importing server/activity,
which eagerly loads the Postgres pool and fails in CI without DATABASE_URL.
Move the helpers into activityFields.ts so npm test stays env-free.
